Update Dispose() and XNAPanel
1. Updated certain controls to dispose unmanaged resources. All controls
now call Dispose() automatically when they are Close()'d. (or, Dispose()
can be called on them individually).

2. XNAPanel now conforms to the conventions from the rest of the
XNAControls library. It uses XNAControl's child model to store a
collection of other XNAControls, and does NOT expose this publicly.
Visible will have to be set true/false in order to show the panel. In
other words, it actually behaves like a control now.

3. Removed EncapsulatingGame as a property of XNAControl since it
duplicated the Game property that already existed.