Enhancements to Visible; bugfixes
There were a couple bugs with the code I submitted. In addition, I added
a new Visible property to XNAControl that updates visibility of any
child controls.

I also updated XNAPanel.cs to include legacy support for
XNAComponentPanel, which will support Components and work as XNAPanel
used to in DepthFirstSearchGame.