ShouldUpdate method
Previously, each Update override checked the same set of criteria (with the exception of XNADialog) to see if it should be updated. I've moved all that into a virtual base class method ShouldUpdate() that can be called to determine whether or not a control should update. Also added in an exception for dialogs that will allow certain dialog types, if they are the top dialog, to still allow other controls to update. This must be set per control.